100 float

+++ Q2/gillespie.cc
@@ -69,13 +69,13 @@ int main (int argc, char** argv) {
     curandCreateGenerator(&gen, CURAND_RNG_PSEUDO_DEFAULT);
 
     float* state;
-    int* dev_X;
+    float* dev_X;
     float* dev_timestep;
 
     cudaMalloc((void**)&state, N * sizeof(float));
     cudaMemset(state, 0, N * sizeof(float));
-    cudaMalloc((void**)&dev_X, N * sizeof(int));
-    cudaMemset(dev_X, 0, N * sizeof(int));
+    cudaMalloc((void**)&dev_X, N * sizeof(float));
+    cudaMemset(dev_X, 0, N * sizeof(float));
     cudaMalloc((void**)&dev_timestep, N * sizeof(float));
     cudaMemset(dev_timestep, 0, N * sizeof(float));
 
@@ -96,16 +96,16 @@ int main (int argc, char** argv) {
     const int T = 1000; // the total time interval after resampling
 
     // the matrix for resampled data
-    int* resamp_X = new int[N * T]();
+    float* resamp_X = new float[N * T]();
 
-    int* dev_resample_X;
-    cudaMalloc((void**)&dev_resample_X, N * T * sizeof(int));
-    cudaMemset(dev_resample_X, 0, N * T * sizeof(int));
+    float* dev_resample_X;
+    cudaMalloc((void**)&dev_resample_X, N * T * sizeof(float));
+    cudaMemset(dev_resample_X, 0, N * T * sizeof(float));
     // the matrix to mark if a time point has been ipdated
     int* is_resampled = new int[N * T]();
     int* dev_is_resampled;
     cudaMalloc((void**)&dev_is_resampled, N * T * sizeof(int));
-    cudaMemset(dev_is_resampled, 0, N * T * sizeof(int));
+    cudaMemset(dev_is_resampled, 0, N * T * sizeof(float));
 
     const float final_time = 100;
     curandSetPseudoRandomGeneratorSeed(gen, 1234);
@@ -189,12 +189,12 @@ int main (int argc, char** argv) {
 
     free(test_accu);
 
-    cudaMemcpy(resamp_X, dev_resample_X, N*T*sizeof(int), cudaMemcpyDeviceToHost);
+    cudaMemcpy(resamp_X, dev_resample_X, N*T*sizeof(float), cudaMemcpyDeviceToHost);
     FILE *total_resample_file = fopen("resample.txt", "w");
 
     for (int i = 0; i < N; ++i) {
         for (int j = 0; j < T; ++j) {
-            fprintf(total_resample_file, "%d ", resamp_X[i*T+j]);
+            fprintf(total_resample_file, "%f ", resamp_X[i*T+j]);
         }
         fprintf(total_resample_file, "\n");
     }
@@ -206,7 +206,7 @@ int main (int argc, char** argv) {
     float* mean = new float[T]();
     for (int i = 0; i < T; ++i) {
         for (int j = 0; j < N; ++j) {
-            mean[i] += (float)resamp_X[j * T + i];
+            mean[i] += resamp_X[j * T + i];
         }
         mean[i] /= N;
     }
